requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
slowapi
python-dotenv
pydantic-settings
pyjwt
passlib
bcrypt==4.0.1
python-multipart
email-validator
fastapi-mail
PyJWT
redis
celery